refactor(capabilities): :recycle: consolidate blog-stats lib + rename GA env vars
Address PR review comments:
- merge src/lib/stats + src/lib/analytics into a single src/lib/blog-stats
  folder (mirrors the /blog/stats route and the blog-stats component), updating
  all import paths
- rename the GA env vars to the GOOGLE_ANALYTICS_ prefix, matching the existing
  GOOGLE_ANALYTICS_API_SECRET convention:
  GA_PROPERTY_ID -> GOOGLE_ANALYTICS_PROPERTY_ID
  GA_SA_CLIENT_EMAIL -> GOOGLE_ANALYTICS_SA_CLIENT_EMAIL
  GA_SA_PRIVATE_KEY -> GOOGLE_ANALYTICS_SA_PRIVATE_KEY
- change the scheduled rebuild from daily to weekly (Mondays 06:00 UTC)
