+++ .github/workflows/scheduled-rebuild.yml
@@ -2,7 +2,7 @@ name: Scheduled Rebuild
 
 on:
   schedule:
-    - cron: '0 6 * * *'
+    - cron: '0 6 * * 1'
   workflow_dispatch:
 
 permissions:
