+++ src/lib/blog-stats/analytics-config.test.ts
@@ -9,8 +9,8 @@ describe("readAnalyticsConfig", () => {
     it("returns null when the property id is missing", () => {
         expect(
             readAnalyticsConfig({
-                GA_SA_CLIENT_EMAIL: "sa@example.com",
-                GA_SA_PRIVATE_KEY: "key",
+                GOOGLE_ANALYTICS_SA_CLIENT_EMAIL: "sa@example.com",
+                GOOGLE_ANALYTICS_SA_PRIVATE_KEY: "key",
             }),
         ).toBeNull();
     });
@@ -18,8 +18,8 @@ describe("readAnalyticsConfig", () => {
     it("returns null when the client email is missing", () => {
         expect(
             readAnalyticsConfig({
-                GA_PROPERTY_ID: "123456",
-                GA_SA_PRIVATE_KEY: "key",
+                GOOGLE_ANALYTICS_PROPERTY_ID: "123456",
+                GOOGLE_ANALYTICS_SA_PRIVATE_KEY: "key",
             }),
         ).toBeNull();
     });
@@ -27,8 +27,8 @@ describe("readAnalyticsConfig", () => {
     it("returns null when the private key is missing", () => {
         expect(
             readAnalyticsConfig({
-                GA_PROPERTY_ID: "123456",
-                GA_SA_CLIENT_EMAIL: "sa@example.com",
+                GOOGLE_ANALYTICS_PROPERTY_ID: "123456",
+                GOOGLE_ANALYTICS_SA_CLIENT_EMAIL: "sa@example.com",
             }),
         ).toBeNull();
     });
@@ -36,9 +36,9 @@ describe("readAnalyticsConfig", () => {
     it("returns a config when all three env vars are present", () => {
         expect(
             readAnalyticsConfig({
-                GA_PROPERTY_ID: "123456",
-                GA_SA_CLIENT_EMAIL: "sa@example.com",
-                GA_SA_PRIVATE_KEY: "-----BEGIN PRIVATE KEY-----\\nabc\\n-----END PRIVATE KEY-----",
+                GOOGLE_ANALYTICS_PROPERTY_ID: "123456",
+                GOOGLE_ANALYTICS_SA_CLIENT_EMAIL: "sa@example.com",
+                GOOGLE_ANALYTICS_SA_PRIVATE_KEY: "-----BEGIN PRIVATE KEY-----\\nabc\\n-----END PRIVATE KEY-----",
             }),
         ).toEqual({
             propertyId: "123456",
